+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [Unreleased]
+### Fixed
+- fixed phone number validation on block-based Checkout page
+- fixed so that the settings page is only loaded when needed
+
 ## [1.20.4]
 ### Fixed
 - fixed so that the shipping method class would be loaded earlier